only use a single bootcache for all environments
We would need some kind of global identifier before loading the app class to
determine a more specific cache. Or the app's index.php would have to do the
loading, but a) would this mean a split of the master.php and b) do we want to
have a single index.php for all apps.

--HG--
branch : multi-backends